+ moved clauses in their own package + moved some functions from RandomUtil to RegexUtil (they seemed better there) + because of ^ some static method references had to be updated
