+++ src/global/BaseRandomTest.java
@@ -1,7 +1,7 @@
 package global;
 
 import global.variables.Clause;
-import global.variables.RandomClause;
+import global.variables.clauses.RandomClause;
 import org.junit.jupiter.api.*;
 
 import java.util.regex.Matcher;
